TinyMCE per block: API refactor
 - extract commands (action creators ?)
 - Handle updating the editor content from state (introducing externalChange)
