+++ tinymce-per-block/src/assets/stylesheets/main.scss
@@ -8,8 +8,8 @@
 @import '~blocks/quote-block/style';
 @import '~blocks/embed-block/style';
 @import '~external/dashicons/style';
-@import '~renderers/block/block-list/style';
-@import '~renderers/html/html-editor/style';
+@import '~renderers/block/style';
+@import '~renderers/html/style';
 @import '~controls/block-arrangement/style';
 @import '~controls/editable-format-toolbar/style';
 @import '~controls/transform-block-toolbar/style';
