+++ tinymce-per-block/src/renderers/block/_style.scss
