--- tinymce-per-block/src/renderers/html/html-editor/index.js
@@ -1,12 +1,0 @@
-/**
- * External dependencies
- */
-import { createElement } from 'wp-elements';
-
-export default function HtmlEditor( { content, onChange } ) {
-	return (
-		<textarea
-			onChange={ ( event ) => onChange( event.target.value ) }
-			className="html-editor" value={ content } />
-	);
-}
